+++ python/benchmarks/bench_contact_3D.py
@@ -7,12 +7,14 @@
 # Multi point constraint problem for linear elasticity with slip conditions
 # between two cubes.
 
+import warnings
 from argparse import ArgumentDefaultsHelpFormatter, ArgumentParser
 from pathlib import Path
 
 import basix.ufl
 import numpy as np
-from dolfinx import default_scalar_type, default_real_type
+from basix.ufl import element
+from dolfinx import default_real_type, default_scalar_type
 from dolfinx.common import Timer, TimingType, list_timings, timing
 from dolfinx.cpp.mesh import entities_to_geometry
 from dolfinx.fem import (Constant, Function, dirichletbc, form, functionspace,
@@ -21,16 +23,16 @@ from dolfinx.io import XDMFFile
 from dolfinx.mesh import (CellType, compute_midpoints, create_mesh,
                           create_unit_cube, locate_entities_boundary, meshtags,
                           refine)
+from dolfinx_mpc.utils import (create_normal_approximation, log_info,
+                               rigid_motions_nullspace, rotation_matrix)
 from mpi4py import MPI
 from petsc4py import PETSc
-from ufl import (Cell, Identity, Mesh, TestFunction, TrialFunction,
-                 VectorElement, dx, grad, inner, sym, tr)
+from ufl import (Identity, Mesh, TestFunction, TrialFunction, dx, grad, inner,
+                 sym, tr)
 
 from dolfinx_mpc import (MultiPointConstraint, apply_lifting, assemble_matrix,
                          assemble_vector)
-from dolfinx_mpc.utils import (create_normal_approximation, log_info,
-                               rigid_motions_nullspace, rotation_matrix)
-import warnings
+
 comm = MPI.COMM_WORLD
 
 if default_real_type == np.float32:
@@ -88,8 +90,7 @@ def mesh_3D_dolfin(theta=0, ct=CellType.tetrahedron, ext="tetrahedron", num_refi
         # Concatenate points and cells
         points = np.vstack([mesh0.geometry.x, mesh1.geometry.x])
         cells = np.vstack([cells0, cells1])
-        cell = Cell(ext, geometric_dimension=points.shape[1])
-        domain = Mesh(VectorElement("Lagrange", cell, 1))
+        domain = Mesh(element("Lagrange", ct.name, 1, shape=(points.shape[1],), gdim=points.shape[1]))
         # Rotate mesh
         points = np.dot(r_matrix, points.T).T.astype(default_real_type)